cli: Restreins la copie et le déplacement de fichiers temporaires

+++ cli/src/io/TmpFile.cpp
@@ -1,6 +1,7 @@
 #include "TmpFile.h"
 
 #include <cstdlib>
+#include <cstring>
 #include <system_error>
 #include <unistd.h>
 
@@ -13,12 +14,28 @@ namespace io {
         }
     }
 
+    TmpFile::TmpFile(TmpFile &&other) noexcept : fd{std::exchange(other.fd, -1)}, filePath{std::move(other.filePath)} {
+        std::memcpy(filename, other.filename, sizeof(TmpFile::filename));
+    }
+
+    TmpFile &TmpFile::operator=(TmpFile &&other) noexcept {
+        if (this == &other) {
+            return *this;
+        }
+        std::memcpy(filename, other.filename, sizeof(TmpFile::filename));
+        fd = std::exchange(other.fd, -1);
+        filePath = std::move(other.filePath);
+        return *this;
+    }
+
     const fs::path &TmpFile::path() const {
         return filePath;
     }
 
     TmpFile::~TmpFile() {
-        unlink(filename);
+        if (fd != -1) {
+            unlink(filename);
+        }
     }
 
     TmpDir::TmpDir() : dirname{"/tmp/oki-XXXXXX"} {
@@ -28,11 +45,28 @@ namespace io {
         dirPath = fs::path{dirname};
     }
 
+    TmpDir::TmpDir(TmpDir &&other) noexcept : dirPath{std::move(other.dirPath)} {
+        std::memcpy(dirname, other.dirname, sizeof(TmpDir::dirname));
+        *other.dirname = '\0';
+    }
+
+    TmpDir &TmpDir::operator=(TmpDir &&other) noexcept {
+        if (this == &other) {
+            return *this;
+        }
+        std::memcpy(dirname, other.dirname, sizeof(TmpDir::dirname));
+        *other.dirname = '\0';
+        dirPath = std::move(other.dirPath);
+        return *this;
+    }
+
     const fs::path &TmpDir::path() const {
         return dirPath;
     }
 
     TmpDir::~TmpDir() {
-        fs::remove_all(dirPath);
+        if (*dirname) {
+            fs::remove_all(dirPath);
+        }
     }
 }
